clean up app file
